add test-d typecheck test
